+++ vscode/src/search/FullTextSearch.test.ts
@@ -1,7 +1,7 @@
+import { SimpleFile } from '../utils/IFile';
 import { IFileSystem } from '../utils/IFileSystem';
 import { InMemFileSystem } from '../utils/InMemFileSystem';
 import { IFullTextSearch } from './IFullTextSearch';
-import { LunrDualFts } from './lunrDualFts';
 import { MyFts } from './myFts';
 
 declare global {
@@ -29,38 +29,29 @@ expect.extend({
 
 const aTextFilePath = '/a/b/c.txt';
 
-class FileAndTags {
-  constructor(
-    public path: string,
-    public text: string,
-    public tags: string[] = []
-  ) {}
-}
-
 let fakeFs: InMemFileSystem;
 
 describe.each([
-  ['LunrDual', (fs: IFileSystem) => new LunrDualFts(fs)],
   ['MyFts', (fs: IFileSystem) => new MyFts(fs, "")]
 ])('full text search: %s', (name, buildFts) => {
   let fts: IFullTextSearch;
 
-  const index = async (files: FileAndTags[]) => {
+  const index = async (files: SimpleFile[]) => {
     fts = buildFts(fakeFs);
     for (const file of files) {
-      await fts.addFile(file.path, file.text, file.tags);
-      fakeFs.writeFile(file.path, file.text);
+      await fts.addFile(file.path(), file.text());
+      fakeFs.writeFile(file.path(), file.text());
     }
   };
 
-  const searchFor = async (query: string, text: string, tags: string[] = []) => {
-    await index([new FileAndTags(aTextFilePath, text, tags)]);
+  const searchFor = async (query: string, text: string) => {
+    await index([new SimpleFile(aTextFilePath, text)]);
     return fts.search(query);
   };
 
-  const modifyFile = async (file: FileAndTags) => {
-    fakeFs.writeFile(file.path, file.text);
-    await fts.onFileModified(file.path, file.text, file.tags);
+  const modifyFile = async (file: SimpleFile) => {
+    fakeFs.writeFile(file.path(), file.text());
+    await fts.onFileModified(file.path(), file.text());
   };
 
   const deleteFile = async (path: string) => {
@@ -75,8 +66,8 @@ describe.each([
 
   it('index and search example', async () => {
     await index([
-      new FileAndTags('blah.txt', 'blah blah stuff'),
-      new FileAndTags('shoe.log', 'shoes and stuff'),
+      new SimpleFile('blah.txt', 'blah blah stuff'),
+      new SimpleFile('shoe.log', 'shoes and stuff'),
     ]);
 
     expect(await fts.search('blah')).toEqual(['blah.txt']);
@@ -115,26 +106,26 @@ describe.each([
 
   it('updates index after file is modified', async () => {
     await index([
-      new FileAndTags('a/b.txt', 'blah blah some stuff and things'),
-      new FileAndTags('a/b/c.log', 'what about shoes and biscuits'),
+      new SimpleFile('a/b.txt', 'blah blah some stuff and things'),
+      new SimpleFile('a/b/c.log', 'what about shoes and biscuits'),
     ]);
 
     let results = await fts.search('blah');
     expect(results.length).toBe(1);
 
-    let modified = new FileAndTags('a/b.txt', 'some stuff and things and more things');
+    let modified = new SimpleFile('a/b.txt', 'some stuff and things and more things');
     await modifyFile(modified);
 
     results = await fts.search('blah');
     expect(results.length).toBe(0);
 
-    modified = new FileAndTags('a/b.txt', 'ok blah is back');
+    modified = new SimpleFile('a/b.txt', 'ok blah is back');
     await modifyFile(modified);
 
     results = await fts.search('blah');
     expect(results.length).toBe(1);
 
-    modified = new FileAndTags('a/b/c.log', 'blah now both notes contain blah');
+    modified = new SimpleFile('a/b/c.log', 'blah now both notes contain blah');
     await modifyFile(modified);
 
     results = await fts.search('blah');
@@ -143,9 +134,9 @@ describe.each([
 
   it('orders search results by relevance', async () => {
     await index([
-      new FileAndTags('lots.of.blah.txt', 'blah blah blah'),
-      new FileAndTags('one.blah.md', 'blah'),
-      new FileAndTags('medium.blah.log', 'blah blah'),
+      new SimpleFile('lots.of.blah.txt', 'blah blah blah'),
+      new SimpleFile('one.blah.md', 'blah'),
+      new SimpleFile('medium.blah.log', 'blah blah'),
     ]);
 
     let results = await fts.search('blah');
@@ -159,12 +150,12 @@ describe.each([
 
   it('orders search results by relevance, after modification', async () => {
     await index([
-      new FileAndTags('lots.of.blah.txt', 'blah blah blah'),
-      new FileAndTags('medium.blah.log', 'blah blah'),
-      new FileAndTags('one.blah.md', 'blah'),
+      new SimpleFile('lots.of.blah.txt', 'blah blah blah'),
+      new SimpleFile('medium.blah.log', 'blah blah'),
+      new SimpleFile('one.blah.md', 'blah'),
     ]);
 
-    let modified = new FileAndTags('one.blah.md', 'turnip');
+    let modified = new SimpleFile('one.blah.md', 'turnip');
     await modifyFile(modified);
 
     const results = await fts.search('blah turnip');
@@ -178,8 +169,8 @@ describe.each([
 
   it('finds file after it was deleted and recreated', async () => {
     await index([
-      new FileAndTags('a/b.txt', 'blah blah some stuff and things'),
-      new FileAndTags('a/b/c.log', 'what about shoes and biscuits'),
+      new SimpleFile('a/b.txt', 'blah blah some stuff and things'),
+      new SimpleFile('a/b/c.log', 'what about shoes and biscuits'),
     ]);
 
     await deleteFile('a/b.txt');
@@ -187,7 +178,7 @@ describe.each([
     let results = await fts.search('blah');
     expect(results).toHaveLength(0);
 
-    const modified = new FileAndTags('a/b.txt', 'ok blah is back');
+    const modified = new SimpleFile('a/b.txt', 'ok blah is back');
     await modifyFile(modified);
 
     results = await fts.search('blah');
@@ -196,11 +187,11 @@ describe.each([
 
   it('does not find file after it was modified then deleted', async () => {
     await index([
-      new FileAndTags('a/b.txt', 'blah blah some stuff and things'),
-      new FileAndTags('a/b/c.log', 'what about shoes and biscuits'),
+      new SimpleFile('a/b.txt', 'blah blah some stuff and things'),
+      new SimpleFile('a/b/c.log', 'what about shoes and biscuits'),
     ]);
 
-    const modified = new FileAndTags('a/b.txt', 'modified blah blah');
+    const modified = new SimpleFile('a/b.txt', 'modified blah blah');
     await modifyFile(modified);
     await deleteFile('a/b.txt');
 
@@ -264,45 +255,6 @@ describe.each([
 
     it('does not match substrings-', async () => {
       await expect(searchFor("beach -ham", "beach nottingham")).toBeFound();
-    });
-  });
-
-  // lunr doesn't support exact phrase matching: https://github.com/olivernn/lunr.js/issues/62
-  // MyFts also doesn't. It could, but meh
-  // describe('phrases', () => {
-  //   it('does not support phrases', async () => {
-  //     await expect(searchFor('"ham is good"', "the ham is good")).not.toBeFound();
-  //   });
-  // });
-
-  describe('search with tags', () => {
-    // MyFts doesn't support tags.
-    const itif = name === "LunrDual" ? it : it.skip;
-
-    itif('finds single tag', async () => {
-      await expect(searchFor("#beef", "The tags are", ['beef', 'chowder'])).toBeFound();
-    });
-
-    itif('finds multiple tags', async () => {
-      await expect(searchFor("#beef #chowder", "The tags are", ['beef', 'chowder'])).toBeFound();
-    });
-
-    itif('does not find missing tag', async () => {
-      await expect(searchFor("#asdf", "The tags are", ['beef', 'chowder'])).not.toBeFound();
-    });
-
-    itif('does not find non tag', async () => {
-      await expect(searchFor("#tags", "The tags are", ['beef', 'chowder'])).not.toBeFound();
-    });
-
-    itif('works with operators', async () => {
-      await expect(searchFor("#beef -#chowder", "The tags are", ['beef', 'chowder'])).not.toBeFound();
-    });
-
-    itif('supports hyphenated tags', async () => {
-      await expect(searchFor("#meat-pie", "I want a", ['meat-pie'])).toBeFound();
-      await expect(searchFor("#meat-pie", "I want a", ['meat'])).not.toBeFound();
-      await expect(searchFor("#meat", "I want a", ['meat-pie'])).not.toBeFound();
     });
   });